% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_duration.R
\name{animate_duration}
\alias{animate_duration}
\title{Title}
\usage{
animate_duration(
  x,
  y,
  t,
  dt = 7,
  Rj = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  gridLonLat = TRUE,
  width = 600,
  height = 400,
  delay = 20,
  title = ""
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
