% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_points.R
\name{animate_points}
\alias{animate_points}
\title{Title}
\usage{
animate_points(
  x,
  y,
  t,
  Rj = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interval = c("day", "week", "month"),
  gridLonLat = TRUE,
  width = 600,
  height = 400,
  delay = 50,
  title = ""
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
