% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_polygons.R
\name{animate_polygons}
\alias{animate_polygons}
\title{Title}
\usage{
animate_polygons(
  region,
  t,
  sf,
  interval = c("day", "week", "month"),
  breaks = NULL,
  gridLonLat = TRUE,
  width = 600,
  height = 400,
  delay = 50
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
