% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnd_modify.R
\name{bnd_modify}
\alias{bnd_modify}
\title{Title}
\usage{
bnd_modify(sf)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
