% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPoints.R
\name{createPoints}
\alias{createPoints}
\title{Title}
\usage{
createPoints(XY.data, crs = NULL, basemap = NULL)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
