% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLongLat.R
\name{distLongLat}
\alias{distLongLat}
\title{Title}
\usage{
distLongLat(lon, lat)
}
\arguments{
\item{lon}{mean.}

\item{lat}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
