% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiTransLibrary.R
\name{EpiTransLibrary}
\alias{EpiTransLibrary}
\title{A Self-made Median Function 2}
\usage{
EpiTransLibrary()
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
