% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLocater.R
\name{GeoLocater}
\alias{GeoLocater}
\title{Calculate the number of cases.}
\usage{
GeoLocater(TimeDF, PointsDF)
}
\arguments{
\item{TimeDF}{Data frame of points regarding time;must contain a column of place (region, polygon name) called 'region', and a column of time called 'time'.}

\item{PointsDF}{Data frame (regarding space) of population points or random points in polygons; must contain a column of place (region, polygon name) called 'region', and a column of x coordinate called 'x', and a column of y coordinate called 'y'.}
}
\value{

}
\description{
Convert polygon data to point coordinates, random sampling and locating to get the cases points' coordinates
}
\examples{
data("EpiTrans")

GeoLocater(TimeDF = covid19,PointsDF = RndPts)
}
