% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdf_GI.R
\name{lpdf_GI}
\alias{lpdf_GI}
\title{Title}
\usage{
lpdf_GI(mean = 20, sd = 9)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
