% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdf_SW.R
\name{lpdf_SW}
\alias{lpdf_SW}
\title{Calculate probability distribution of tranmission distance}
\usage{
lpdf_SW(mean = 125)
}
\arguments{
\item{mean}{Mean tranmission distance of exponential distribution}
}
\value{

}
\description{
Spatial weighting function; probability distribution (exponential distribution) of tranmission distance
}
\examples{
lpdf_SW(mean = 125)
}
