% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.points.R
\name{plot.points}
\alias{plot.points}
\title{Title}
\usage{
\method{plot}{points}(
  x,
  y,
  t = NULL,
  Rj = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interact = TRUE,
  grid = TRUE,
  title = ""
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
