% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Tij.R
\name{plot.Tij}
\alias{plot.Tij}
\title{Title}
\usage{
\method{plot}{Tij}(Tij, sf)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
