% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hex.R
\name{plot_hex}
\alias{plot_hex}
\title{Title}
\usage{
plot_hex(
  x,
  y,
  Rj = NULL,
  grid.n = 50,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interact = TRUE,
  gridLonLat = TRUE,
  title = ""
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
