% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kde.R
\name{plot_kde}
\alias{plot_kde}
\title{Visualizing Point pattern}
\usage{
plot_kde(
  x,
  y,
  grid.n = 50,
  bandwidth = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interact = TRUE,
  gridLonLat = TRUE,
  title = ""
)
}
\arguments{
\item{x}{Vector of x coordinates of data points.}

\item{y}{Vector of y coordinates of data points.}

\item{grid.n}{Number of grid cells in one direction (shorter side).}

\item{bandwidth}{Specifing the band width for KDE.}

\item{crs}{Coordinate reference system of data points: numeric (EPSG code), object of class 'crs', or input string for st_crs; if NULL, using CRS of basemap, or using WGS84 (EPSG:4326).}

\item{bnd}{A 4-length numerical vector of the form c(xmin, ymin, xmax, ymax) which gives the number of x and y coordinates' ranges of the map.}

\item{basemap}{Object of class 'sf'; basemap plotting on map.}

\item{interact}{Logical; if TRUE, plot thematic interactive map; if FALSE, static plotting non-interactive thematic map.}

\item{gridLonLat}{Logical; draw latitude and longitude grid on non-interactive map.}

\item{title}{Main title of plot.}
}
\value{

}
\description{
Plot the spatial pattern of cases by KDE method; smoothed overview of the clustering pattern of an outbreak
}
\examples{
data("EpiTrans")

plot_kde(x = dengue$long, y= dengue$lat, grid.n = 30, basemap = Taiwan)
}
