% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpdf_SW.R
\name{plot_lpdf_SW}
\alias{plot_lpdf_SW}
\title{Visualizing probability density functions.}
\usage{
plot_lpdf_SW(mean = 125, from = 0, to = 1000)
}
\arguments{
\item{mean}{Mean tranmission distance of exponential distribution.}

\item{from}{Minimum of x of the plot.}

\item{to}{Maximum of x of the plot.}
}
\description{
Plot the probability distribution of tranmission distance
}
\examples{
plot_lpdf_SW(mean = 125, from = 0, to = 1000)
}
