% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polygons.R
\name{plot_polygons}
\alias{plot_polygons}
\title{Title}
\usage{
plot_polygons(region, sf, interact = TRUE, gridLonLat = TRUE)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
