% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trans.R
\name{plot_trans}
\alias{plot_trans}
\title{Title}
\usage{
plot_trans(Rj, from = 1:200, to = 1:200, title = NULL)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
