% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceMatrix.R
\name{ReduceMatrix}
\alias{ReduceMatrix}
\title{Title}
\usage{
ReduceMatrix(matrix, group)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
