% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rj.R
\name{Rj}
\alias{Rj}
\title{Calculate reproductive numbers}
\usage{
Rj(
  t,
  x,
  y,
  adjSP = TRUE,
  GI.pdf = lpdf_GI(),
  SW.pdf = lpdf_SW(),
  unit_coord = c("degree", "meter")
)
}
\arguments{
\item{t}{Vector of time of data points.}

\item{x}{Vector of x coordinates of data points.}

\item{y}{Vector of y coordinates of data points.}

\item{adjSP}{Logical; if TRUE, reproductive numbers will be calculated with spatial adjustment; if FALSE, static plotting non-interactive thematic map reproductive numbers will be calculated only considering temporal relation.}

\item{GI.pdf}{Probability distribution of generation interval; it can be generated by function 'lpdf_GI'.}

\item{SW.pdf}{Probability distribution (exponential distribution) of tranmission distance; it can be generated by function 'lpdf_SW'.}

\item{unit_coord}{The unit of the coordination; if 'degree' then calculate distance througth 'distLongLat'; if 'meter' then calculate distance directly.}
}
\value{

}
\description{
Calculate individual reproductive numbers given data (n-length vectors : t, x, y) and prespecified likelihood functions (lpdf_GI, lpdf_SP)
}
\examples{
data("EpiTrans")

res_adj = Rj(t = dengue$date, x = dengue$long, y = dengue$lat,GI.pdf = lpdf_GI(),SW.pdf = lpdf_SW(), adjSP = TRUE)
res = Rj(t = dengue$date, GI.pdf = lpdf_GI(), adjSP = FALSE)
}
