% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialNumbering.R
\name{SequentialNumbering}
\alias{SequentialNumbering}
\title{Title}
\usage{
SequentialNumbering(region)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
