% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_polygons.R
\name{animate_polygons}
\alias{animate_polygons}
\title{Visualizing the spatio-temporal pattern of choropleth map}
\usage{
animate_polygons(
  region,
  t,
  sf,
  interval = c("day", "week", "month"),
  breaks = NULL,
  gridLonLat = TRUE,
  width = 600,
  height = 400,
  delay = 50
)
}
\arguments{
\item{region}{Vector of region (polygon) name of data points.}

\item{t}{Vector of time of data points.}

\item{sf}{Object of class 'sf'; polygon layer of statistical data (region data of these data points).}

\item{interval}{Character; time resolution(day, week, month) of animation ; aggregrate data through day, week or month.}

\item{breaks}{Numeric vector to tell how chorolepth classify.}

\item{gridLonLat}{Logical; draw latitude and longitude grid on non-interactive map.}

\item{width}{Width of the animation file (in pixels).}

\item{height}{Height of the animation file (in pixels).}

\item{delay}{Delay time between images (in 1/100th of a second).}
}
\value{

}
\description{
Generate animation of chorolepth map of cases by grouping region
}
\examples{
data("EpiTrans")
covid = GeoLocater(TimeDF = covid19,PointsDF = RndPts)

animate_polygons(region = covid$region,t = covid$date,sf = Taipei)
}
