% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnd_modify.R
\name{bnd_modify}
\alias{bnd_modify}
\title{Adjust the boundary of the plot map}
\usage{
bnd_modify(sf)
}
\arguments{
\item{sf}{Object of class 'sf'.}
}
\value{

}
\description{
Auto adjust the boundary of the plot map
}
\examples{
data("EpiTrans")

bnd <- bnd_modify(Taiwan)
}
