% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaseMap.R
\name{CreateBaseMap}
\alias{CreateBaseMap}
\title{Title}
\usage{
CreateBaseMap(basemap, bnd, gridLonLat, interact, alphaGrid = 1)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
