% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPoints.R
\name{createPoints}
\alias{createPoints}
\title{Create points data}
\usage{
createPoints(XY.data, crs = NULL, basemap = NULL)
}
\arguments{
\item{XY.data}{Data frame of points;must (at least) contain a column of x coordinate called 'x', and a column of y coordinate called 'y';and other column if needed such as 't' (time), 'Rj' (reproductive numbers ).}

\item{crs}{Coordinate reference system of data points: numeric (EPSG code), object of class 'crs', or input string for st_crs; if NULL, using CRS of basemap, or using WGS84 (EPSG:4326).}

\item{basemap}{Object of class 'sf'; reference crs to create points on the crs of this basemap}
}
\value{

}
\description{
Create sf points data by data frame with coordinates
}
\examples{
data("EpiTrans")

points = RndPts[,2:3]

createPoints(points,crs = 4326,basemap = Taiwan)
}
