% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLongLat.R
\name{distLongLat}
\alias{distLongLat}
\title{Calculate geometry distance}
\usage{
distLongLat(lon, lat)
}
\arguments{
\item{lon}{Vector of longitude; its length stands for the number of points.}

\item{lat}{Vector of latitude; its length stands for the number of points.}
}
\value{

}
\description{
Calculate distance of longtitude and latitude
}
\examples{
data("EpiTrans")

covid = GeoLocater(TimeDF = covid19,PointsDF = RndPts)
lon <- covid$x[1:100]
lat <- covid$y[1:100]

distLongLat(lon,lat)
}
