% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLocater.R
\name{GeoLocater}
\alias{GeoLocater}
\title{Title}
\usage{
GeoLocater(TimeDF, PointsDF)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
