% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdf_GI.R
\name{lpdf_GI}
\alias{lpdf_GI}
\title{Calculate the probability distribution of generation interval}
\usage{
lpdf_GI(mean = 20, sd = 9)
}
\arguments{
\item{mean}{Mean of the probability distribution of generation interval.}

\item{sd}{Standard deviation of the probability distribution of generation interval.}
}
\value{

}
\description{
Probability distribution (gamma distribution) of generation interval, defined as the time interval between the infection (or onset) time of infector and infectee.
}
\examples{
lpdf_GI(mean = 20,sd = 9)
}
