% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi.R
\name{plot_epi}
\alias{plot_epi}
\title{Title}
\usage{
plot_epi(t, interval = c("day", "week", "month"))
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
