% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kde.R
\name{plot_kde}
\alias{plot_kde}
\title{Title}
\usage{
plot_kde(
  x,
  y,
  grid.n = 50,
  bandwidth = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interact = TRUE,
  gridLonLat = TRUE,
  title = ""
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
