% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpdf_GI.R
\name{plot_lpdf_GI}
\alias{plot_lpdf_GI}
\title{Title}
\usage{
plot_lpdf_GI(mean = 20, sd = 9, from = 0, to = 50)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
