% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpdf_SW.R
\name{plot_lpdf_SW}
\alias{plot_lpdf_SW}
\title{Title}
\usage{
plot_lpdf_SW(mean = 125, from = 0, to = 1000)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
