% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Visualizing Point pattern}
\usage{
plot_points(
  x,
  y,
  t = NULL,
  Rj = NULL,
  crs = NULL,
  bnd = NULL,
  basemap = NULL,
  interact = TRUE,
  gridLonLat = TRUE,
  title = ""
)
}
\arguments{
\item{x}{Vector of x coordinates of data points.}

\item{y}{Vector of y coordinates of data points.}

\item{t}{Vector of time of data points.}

\item{Rj}{Vector of individual reproductive numbers of data points.}

\item{crs}{Coordinate reference system of data points: numeric (EPSG code), object of class 'crs', or input string for st_crs; if NULL, using CRS of basemap, or using WGS84 (EPSG:4326).}

\item{bnd}{A 4-length numerical vector of the form c(xmin, ymin, xmax, ymax) which gives the number of x and y coordinates' ranges of the map.}

\item{basemap}{Object of class 'sf'; basemap plotting on map.}

\item{interact}{Logical; if TRUE, plot thematic interactive map; if FALSE, static plotting non-interactive thematic map.}

\item{gridLonLat}{Logical; draw latitude and longitude grid on non-interactive map.}

\item{title}{Main title of plot.}
}
\description{
Plot the spatial points of cases; point pattern is the most direct and exact way to visualize spatial pattern of an outbreak.
}
\examples{
data("EpiTrans")

plot_points(x = dengue$long, y= dengue$lat, crs= NULL,basemap = Taiwan)
}
