% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reigonTrans.R
\name{plot_reigonTrans}
\alias{plot_reigonTrans}
\title{Title}
\usage{
plot_reigonTrans(Pij, region, sf, time = NULL)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
