% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Rt.R
\name{plot_Rt}
\alias{plot_Rt}
\title{Title}
\usage{
plot_Rt(
  t,
  Rt,
  Rt2 = NULL,
  percentile = c(0, 1),
  interval = c("day", "week", "month"),
  col = "orange",
  col2 = "darkgreen",
  label = "Rt",
  label2 = "Rt2"
)
}
\arguments{
\item{mean}{mean.}

\item{sd}{sd}
}
\description{
Descriptions Here
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
