% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rj.R
\name{Rj}
\alias{Rj}
\title{calculate individual R}
\usage{
Rj(
  t,
  x,
  y,
  adjSP = TRUE,
  GI.pdf = lpdf_GI(),
  SW.pdf = lpdf_SW(),
  unit_coord = c("degree", "meter")
)
}
\arguments{
\item{mean}{mean.}
}
\description{
calculate individual R and matrix of transmission probability
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
