% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialNumbering.R
\name{SequentialNumbering}
\alias{SequentialNumbering}
\title{Give sequential number by region}
\usage{
SequentialNumbering(region)
}
\arguments{
\item{region}{Vector of region name of each data points}
}
\value{

}
\description{
Give sequential number by region
}
\examples{
SequentialNumbering(covid19$region)
}
